Рефакторинг кода и изменение шрифтов
- Календарь вынесен в отдельный компонент `CalendarGrid`.
- Заменены шрифты: Ubuntu на Montserrat и Rubik.
- Реализован пример `SwipeToDismissListItems` для демонстрации возможности удаления элементов свайпом.
- Изменен заголовок на экране HomeScreen.
