+++ app/src/main/java/com/wheatley/morph/ui/theme/Type.kt
@@ -2,8 +2,9 @@ package com.wheatley.morph.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.googlefonts.Font
+
 import androidx.compose.ui.text.googlefonts.GoogleFont
+import androidx.compose.ui.text.googlefonts.Font
 import com.wheatley.morph.R
 
 val provider = GoogleFont.Provider(
@@ -14,14 +15,14 @@ val provider = GoogleFont.Provider(
 
 val bodyFontFamily = FontFamily(
     Font(
-        googleFont = GoogleFont("Ubuntu"),
+        googleFont = GoogleFont("Montserrat"),
         fontProvider = provider,
     )
 )
 
 val displayFontFamily = FontFamily(
     Font(
-        googleFont = GoogleFont("Ubuntu"),
+        googleFont = GoogleFont("Rubik"),
         fontProvider = provider,
     )
 )
